Crate storage_api

A proof-of-concept implementation of (one version of) the storages proposal.

Describing the raw storage API, we have:

Storage: a storage that can store objects

SliceStorage: a storage for growable slices

MultipleStorage: a storage that can store multiple objects

SharedMutabilityStorage and PinningStorage

Providing a safe wrapper around Storage use (up to uninit memory):

RawBox: a raw (uninit payload) version of std Box

RawVec: a raw (uninit payload) version of std Vec

Useful implementations of Storage:

InlineStorage: single storage located in the storage’s bytes

AllocStorage: full-featured storage via allocation

SmallStorage: inline storage with a fallback to allocation

Structs

AllocStorage

A storage that stores objects via an Allocator.

InlineStorage

A single storage which stores objects inline.

InlineStorageHandle

A handle into an [InlineStorage].

RawBox

A raw box around some storage. Bundles the storage and its handle.

RawVec

A raw vec around some slice storage. Bundles the storage and its handle.

SmallStorage

A single storage which stores objects inline if it fits, otherwise falling back to using an Allocator.

SmallStorageHandle

Traits

Handle

Types which can be used as a storage handle.

MultipleStorage

A storage that can create multiple handles.

PinningStorage

A storage that creates pinned handles.

SharedMutabilityStorage

A storage that serves as a uniqueness barrier.

SliceStorage

A storage that can reallocate to adjust the length of slice objects.

Storage

Types which can be used to store objects.