Struct storage_api::Box

pub struct Box<T: ?Sized, S: Storage> { /* private fields */ }

A pointer type for heap allocation. A tiny subset of std’s Box.

Implementations

impl<'a, U> Box<U, DynStorage<'a>> where
U: ?Sized + Pointee<Metadata = DynMetadata<U>>,

pub fn boxed<A>(boxed: Box<U, AllocStorage<A>>) -> Self where
A: Copy + Allocator,
Bool<{ is_zst::<A>() }>: True,

Construct a dynamic storage box from a standard box.

pub fn inline(boxed: Box<U, InlineStorage<usize>>) -> Self

Construct a dynamic storage box inline.

pub unsafe fn take(taken: &'a mut ManuallyDrop<U>) -> Self

Construct a dynamic storage box by taking someone else’s allocation.

impl<T: ?Sized, S: Storage> Box<T, S>

pub fn new_in(t: T, storage: S) -> Self where
T: Sized,

pub fn into_raw_parts(this: Self) -> (S::Handle, <T as Pointee>::Metadata, S)

pub unsafe fn from_raw_parts(
handle: S::Handle,
metadata: <T as Pointee>::Metadata,
storage: S
) -> Self