Trait storage_api::SliceStorage

pub unsafe trait SliceStorage: Storage {
    unsafe fn grow<T>(
        &mut self, 
        handle: Self::Handle<[T]>, 
        new_len: usize
    ) -> Result<Self::Handle<[T]>, AllocError>;
    unsafe fn shrink<T>(
        &mut self, 
        handle: Self::Handle<[T]>, 
        new_len: usize
    ) -> Result<Self::Handle<[T]>, AllocError>;
}

A storage that can reallocate to adjust the length of slice objects.

Automatically provided for any MultipleStorage by allocating a new object and copying the old allocation into the new one.

Required methods

unsafe fn grow<T>(
&mut self,
handle: Self::Handle<[T]>,
new_len: usize
) -> Result<Self::Handle<[T]>, AllocError>

Grow a slice handle to a larger size.

If this function succeeds, then the old handle is invalidated and the handled object has been moved into the new handle. The new length is uninitialized.

(Do we want a grow_zeroed?)

If this function fails, then the old handle is not invalidated and still contains the object.

Safety

The handle must have previously been created by this storage, and must not have been destroyed or invalidated.

new_len must be longer than the existing slice length.

unsafe fn shrink<T>(
&mut self,
handle: Self::Handle<[T]>,
new_len: usize
) -> Result<Self::Handle<[T]>, AllocError>

Shrink a slice handle to a smaller size.

If this function succeeds, then the old handle is invalidated and the prefix of the handled object has been moved into the new handle.

If this function fails, then the old handle is not invalidated and still contains the object.

Safety

The handle must have previously been created by this storage, and must not have been destroyed or invalidated.

new_len must be shorter than the existing slice length.