Trait storage_api::MultipleStorage

pub unsafe trait MultipleStorage<T: ?Sized>: Storage<T> { }

A storage that can create multiple handles.

The restriction that create can not be called until the previously created handle has been destroyed is removed. You can create multiple handles that are all valid at the same time. This does not change the requirements of resolve or resolve_mut; only one pointer returned from resolve_mut is valid at a time.

(In theory, we’re missing the ability to create multiple handles to a !SharedMutabilityStorage and access their objects mutably simultaneously. This is possible to do soundly (for the homogeneously-typed case, store a slice/array and split_at_mut it), but seems rare enough in practice that the complexity to support it doesn’t seem warranted? Such a function would live on this trait and could look like the following:

unsafe fn resolve_many_mut<T: ?Sized, N: usize>(
    &mut self,
    handles: [Self::Handle<T>; N],
) -> [NonNull<T>; N];

The reason for this restriction is that reborrowing &mut Storage to call resolve_mut invalidates any references/pointers derived from the previous reborrow of &mut Storage. This is fundamental to &mut being noalias.)