Trait storage_api::Storage

pub unsafe trait Storage<T: ?Sized> {
    type Handle: Handle<T>;
    unsafe fn create(
        &mut self, 
        meta: <T as Pointee>::Metadata
    ) -> Result<Self::Handle, AllocError>;
    unsafe fn destroy(&mut self, handle: Self::Handle);
    unsafe fn resolve_metadata(
        &self, 
        handle: Self::Handle
    ) -> <T as Pointee>::Metadata;
    unsafe fn resolve(&self, handle: Self::Handle) -> NonNull<T>;
    unsafe fn resolve_mut(&mut self, handle: Self::Handle) -> NonNull<T>;
}

Types which can be used to store objects.

The behavior of this trait is refined by traits PinningStorage, MultipleStorage, and SharedMutabilityStorage.

(I’ve used create/destroy to clearly separate this from Allocator’s allocate/deallocate language, but naming is to be bikeshedded further.)

Associated Types

type Handle: Handle<T>

The handle which is used to access

Required methods

unsafe fn create(
&mut self,
meta: <T as Pointee>::Metadata
) -> Result<Self::Handle, AllocError>

Create an object handle in this storage.

The handled object is not initialized.

(Do we want a create_zeroed?)

Safety

Any previously created handles have been destroyed.

(This can maybe be loosened to it invalidating existing handles?)

The metadata must describe a layout valid for a rust object.

This exists due to the safety requirements of size_of_val_raw and align_of_val_raw. I think we need a way to go compute layout from metadata safely, with a check that it produces a valid layout.

On each unsized kind, this would imply:

slices: size computation uses saturating/checked multiplication

traits: vtable must always be valid (as a safety invariant)

composites: size computation uses saturating/checked addition

externs: any valid metadata must compute valid size/align or None

unsafe fn destroy(&mut self, handle: Self::Handle)

Destroy an object handle in this storage.

The handled object is not modified nor required to be valid in any way.

Safety

The handle must have previously been created by this storage, and must not have been destroyed.

unsafe fn resolve_metadata(
&self,
handle: Self::Handle
) -> <T as Pointee>::Metadata

Resolve the metadata of a handle in this storage.

Safety

The handle must have previously been created by this storage, and must not have been destroyed or invalidated.

unsafe fn resolve(&self, handle: Self::Handle) -> NonNull<T>

Resolve an object handle in this storage to a pointer.

The returned pointer is valid for reads only and is invalidated when the storage is moved or used mutably.

Safety

The handle must have previously been created by this storage, and must not have been destroyed or invalidated.

unsafe fn resolve_mut(&mut self, handle: Self::Handle) -> NonNull<T>

Resolve an object handle in this storage to a pointer.

The returned pointer is valid for both reads and writes and is invalidated when the storage is moved or used mutably. (This includes but is not limited to further calls to resolve_mut.)

Safety

The handle must have previously been created by this storage, and must not have been destroyed or invalidated.